# Fitness App/fitness-ui/src/app/services/models/member-wrapper.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface MemberWrapper {
  age?: number;
  bmi?: number;
  height?: number;
  id?: number;
  name?: string;
  weight?: number;
}