# Fitness App/fitness-ui/src/app/services/models/user-wrapper.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface UserWrapper {
  contactNumber?: string;
  email?: string;
  id?: number;
  name?: string;
  status?: string;
}